+++ plone/app/event/tests/test_base_module.py
@@ -103,6 +103,16 @@ class TestBaseModule(unittest.TestCase):
             DateTime('2011/11/11 11:00:00 UTC')
         )
 
+    def test_cal_to_strftime_wkday(self):
+        from plone.app.event.base import cal_to_strftime_wkday
+        li = [cal_to_strftime_wkday(day) for day in range(0,6)]
+        self.assertTrue(li == [1,2,3,4,5,6,0])
+
+    def test_strftime_to_cal_wkday(self):
+        from plone.app.event.base import cal_to_strftime_wkday
+        li = [cal_to_strftime_wkday(day) for day in range(0,6)]
+        self.assertTrue(li == [6,0,1,2,3,4,5])
+
 
 class TestCalendarLinkbase(unittest.TestCase):
     # TODO: test overriding of ICalendarLinkbase